requirements.txt
dnspython==2.2.1
Flask==2.1.2
Flask-Login==0.6.2
flask-restx==0.5.1
Flask-Session==0.4.0
gunicorn==20.1.0
opencv-python==4.6.0.66
numpy==1.23.0
paho-mqtt==1.6.1
pillow==9.2.0
pymongo==4.2.0
python-dotenv==0.21.0
requests==2.28.1
